feat(profile): move agents service file up a folder
