+++ server.js
@@ -3,7 +3,7 @@ const services = [
   require('./tasks/services/plans'),
   require('./tasks/services/works'),
   require('./ordering/services/orders'),
-  require('./agents/services/hooks/agents')
+  require('./agents/services/agents')
 ]
 
 export default {
